order processing

+++ SimpleInventory.Core/Models/ItemModel.cs
@@ -18,8 +18,8 @@ namespace SimpleInventory.Core.Models
         public string? Name { get; set; }
         public string? Description { get; set; }
         public string? Type { get; set; }
-        public int Quantity { get; set; }
-        public double Price { get; set; }
+        public uint Quantity { get; set; }
+        public decimal Price { get; set; }
 
         public ItemModel()
         {
